# packages/backend/src/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/backups/{urn}/backup": {
      "post": {
        "operationId": "backupApp",
        "parameters": [
          {
            "name": "urn",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BackupRequestDto"
                }
              }
            }
          }
        },
        "tags": [
          "Backups"
        ]
      }
    },
    "/api/backups/{urn}/restore": {
      "post": {
        "operationId": "restoreAppBackup",
        "parameters": [
          {
            "name": "urn",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RestoreAppBackupDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BackupRequestDto"
                }
              }
            }
          }
        },
        "tags": [
          "Backups"
        ]
      }
    },
    "/api/backups/{urn}": {
      "get": {
        "operationId": "getAppBackups",
        "parameters": [
          {
            "name": "urn",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetAppBackupsDto"
                }
              }
            }
          }
        },
        "tags": [
          "Backups"
        ]
      },
      "delete": {
        "operationId": "deleteAppBackup",
        "parameters": [
          {
            "name": "urn",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteAppBackupBodyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Backups"
        ]
      }
    },
...
  },
  "info": {
    "title": "Runtipi API",
    "description": "API specs for Runtipi",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "BackupRequestDto": {
        "type": "object",
        "properties": {
          "requestId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "requestId"
        ]
      },
      "RestoreAppBackupDto": {
        "type": "object",
        "properties": {
          "filename": {
            "type": "string"
          }
        },
        "required": [
          "filename"
        ]
      },
      "GetAppBackupsDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "size": {
                  "type": "number"
                },
                "date": {
                  "type": "number"
                }
              },
              "required": [
                "id",
                "size",
                "date"
              ]
            }
          },
          "total": {
            "type": "number"
          },
          "currentPage": {
            "type": "number"
          },
          "lastPage": {
            "type": "number"
          }
        },
        "required": [
          "data",
          "total",
          "currentPage",
          "lastPage"
        ]
      },
      "DeleteAppBackupBodyDto": {
        "type": "object",
        "properties": {
          "filename": {
            "type": "string"
          }
        },
        "required": [
          "filename"
        ]
      },
...
    }
  }
}